修复异地登录和投注记录查询的bug以及添加代理首页的接口和代理添加会员功能

+++ tower-model/src/main/java/com/tower/dto/AgentDto.java
@@ -33,8 +33,8 @@ public class AgentDto implements Serializable {
     private Integer newNum;
 
     @ApiModelProperty(value = "历史总奖励")
-    private Integer totalAward;
+    private BigDecimal totalAward;
 
     @ApiModelProperty(value = "可提现金额")
-    private Integer canAward;
+    private BigDecimal canAward;
 }
